refactor(config): clarify package root export policy
