+++ docs/api.md
@@ -16,183 +16,7 @@ Analysis module for Frame Compare.
 
 ## frame_compare.config
 
-Convenience config re-export surface for Frame Compare.
-
-### AnalysisConfig
-
-`AnalysisConfig`
-
-Frame selection and analysis settings.
-
-### apply_cli_overrides
-
-`apply_cli_overrides(config: ConfigSchema, cli_args: Mapping[str, object]) -> ConfigSchema`
-
-Apply CLI arguments as config overrides.
-
-### apply_preset
-
-`apply_preset(config: ConfigSchema, preset_name: str, presets_dir: Path | None = None) -> ConfigSchema`
-
-Apply preset overrides to config.
-
-### AudioAlignmentConfig
-
-`AudioAlignmentConfig`
-
-Audio alignment and interactive alignment configuration.
-
-### CLI_OVERRIDE_MAP
-
-`CLI_OVERRIDE_MAP` — constant (dict)
-
-### ColorConfig
-
-`ColorConfig`
-
-Tonemapping configuration for HDR sources.
-
-### ConfigSchema
-
-`ConfigSchema`
-
-Root configuration schema using pydantic-settings.
-
-### DEFAULT_CONFIG_TOML
-
-`DEFAULT_CONFIG_TOML` — constant (str)
-
-### DiagnosticsConfig
-
-`DiagnosticsConfig`
-
-Optional diagnostic outputs for development and debugging.
-
-### DoviConfig
-
-`DoviConfig`
-
-Dolby Vision metadata extraction configuration.
-
-### get_default_config
-
-`get_default_config() -> ConfigSchema`
-
-Get config with all default values (no TOML, no env).
-
-### list_presets
-
-`list_presets(presets_dir: Path | None = None) -> list[str]`
-
-List available preset names (sorted case-insensitive).
-
-### load_config
-
-`load_config(config_path: Path | None = None, overrides: dict[str, object] | None = None) -> ConfigSchema`
-
-Load configuration from TOML file with overrides.
-
-### load_config_from_env
-
-`load_config_from_env() -> ConfigSchema`
-
-Load config from environment variables only (no TOML file).
-
-### load_preset
-
-`load_preset(name: str, presets_dir: Path | None = None) -> dict[str, object]`
-
-Load preset data by name.
-
-### LogFormat
-
-`LogFormat`
-
-Logging output format.
-
-### LoggingConfig
-
-`LoggingConfig`
-
-Logging configuration (level, format, optional file path).
-
-### LogLevel
-
-`LogLevel`
-
-Log level using uppercase values per Python logging convention.
-
-### OverlayMode
-
-`OverlayMode`
-
-Overlay verbosity level for rendered screenshots.
-
-### PathsConfig
-
-`PathsConfig`
-
-Filesystem paths relative to the workspace root.
-
-### ReportConfig
-
-`ReportConfig`
-
-HTML report generation configuration.
-
-### save_preset
-
-`save_preset(name: str, config: ConfigSchema, presets_dir: Path | None = None) -> Path`
-
-Save current config as preset.
-
-### ScreenshotsConfig
-
-`ScreenshotsConfig`
-
-Screenshot rendering settings (encoder choice, overlays, compression).
-
-### SelectionMode
-
-`SelectionMode`
-
-Frame selection strategy for choosing representative comparison frames.
-
-### SlowpicsConfig
-
-`SlowpicsConfig`
-
-slow.pics upload configuration and retry policy.
-
-### TmdbConfig
-
-`TmdbConfig`
-
-TMDB metadata lookup configuration.
-
-### ToneCurve
-
-`ToneCurve`
-
-Tone curve algorithm used by the tonemap implementation.
-
-### TonemapPreset
-
-`TonemapPreset`
-
-Named tonemap preset applied when tonemapping HDR to SDR.
-
-### ViewerMode
-
-`ViewerMode`
-
-Default comparison viewer mode for generated HTML reports.
-
-### Visibility
-
-`Visibility`
-
-slow.pics gallery visibility setting.
+Configuration subsystem namespace root.
 
 ## frame_compare.orchestration
