+++ src/frame_compare/config/__init__.py
@@ -1,75 +1,7 @@
-"""Convenience config re-export surface for Frame Compare.
+"""Configuration subsystem namespace root.
 
-Internal owners should prefer direct imports from the concrete config modules
-(`loader`, `overrides`, `presets`, and `schema`) when that keeps dependency
-ownership clearer.
+To keep dependency ownership explicit, all consumers should import directly
+from the concrete modules: schema, loader, overrides, and presets.
 """
 
-from frame_compare.config.defaults import DEFAULT_CONFIG_TOML
-from frame_compare.config.loader import (
-    get_default_config,
-    load_config,
-    load_config_from_env,
-)
-from frame_compare.config.overrides import CLI_OVERRIDE_MAP, apply_cli_overrides
-from frame_compare.config.presets import (
-    apply_preset,
-    list_presets,
-    load_preset,
-    save_preset,
-)
-from frame_compare.config.schema import (
-    AnalysisConfig,
-    AudioAlignmentConfig,
-    ColorConfig,
-    ConfigSchema,
-    DiagnosticsConfig,
-    DoviConfig,
-    LogFormat,
-    LoggingConfig,
-    LogLevel,
-    OverlayMode,
-    PathsConfig,
-    ReportConfig,
-    ScreenshotsConfig,
-    SelectionMode,
-    SlowpicsConfig,
-    TmdbConfig,
-    ToneCurve,
-    TonemapPreset,
-    ViewerMode,
-    Visibility,
-)
-
-__all__ = [
-    "load_config",
-    "load_config_from_env",
-    "get_default_config",
-    "ConfigSchema",
-    "CLI_OVERRIDE_MAP",
-    "apply_cli_overrides",
-    "apply_preset",
-    "list_presets",
-    "load_preset",
-    "save_preset",
-    "DEFAULT_CONFIG_TOML",
-    "SelectionMode",
-    "OverlayMode",
-    "TonemapPreset",
-    "ToneCurve",
-    "Visibility",
-    "ViewerMode",
-    "LogLevel",
-    "LogFormat",
-    "PathsConfig",
-    "AnalysisConfig",
-    "AudioAlignmentConfig",
-    "ScreenshotsConfig",
-    "ColorConfig",
-    "SlowpicsConfig",
-    "TmdbConfig",
-    "ReportConfig",
-    "DoviConfig",
-    "DiagnosticsConfig",
-    "LoggingConfig",
-]
+__all__ = []
