+++ tests/cli/test_cli_commands.py
@@ -14,8 +14,8 @@ from typer.main import get_command
 from typer.testing import CliRunner
 
 from frame_compare.cli.entry import _maybe_open_report, _stabilize_typer_help_width, app
-from frame_compare.config import OverlayMode, ToneCurve, TonemapPreset
 from frame_compare.config.loader import get_default_config
+from frame_compare.config.schema import OverlayMode, ToneCurve, TonemapPreset
 from frame_compare.errors import (
     ConfigNotFoundError,
     ErrorContext,
