+++ tests/orchestration/test_ffmpeg_overlay_application.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import pytest
 from PIL import Image
 
-from frame_compare.config import ConfigSchema
+from frame_compare.config.schema import ConfigSchema
 from frame_compare.orchestration.context import (
     ClipFingerprint,
     ClipProbeSnapshot,
